ultimofix

+++ tiendaweb/src/main/java/com/tiendavideojuegos/tiendaweb/dao/VideojuegoDaoImpl.java
@@ -14,7 +14,7 @@ public class VideojuegoDaoImpl implements CrudDaoInterface<VideojuegoDto> {
         return null;
     }
 
-    public List<VideojuegoDto> getbyId(float num) {
+    public List<VideojuegoDto> getbyId(int num) {
        
         return null;
     }
